refactor(sdk-coin-xrp): move interfaces to `/lib`
Ticket: WIN-3569

TICKET: WIN-3569
