+++ modules/sdk-coin-xrp/src/index.ts
@@ -1,3 +1,4 @@
 export * from './xrp';
 export * from './txrp';
 export * from './register';
+export * from './lib/iface';
